Modularize flight-sql
