+++ java/flight/flight-sql/src/main/java/module-info.java
@@ -1,0 +1,29 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+module org.apache.arrow.flight.sql {
+  exports org.apache.arrow.flight.sql;
+  exports org.apache.arrow.flight.sql.example;
+  exports org.apache.arrow.flight.sql.util;
+
+  requires com.google.common;
+  requires java.sql;
+  requires org.apache.arrow.flight.core;
+  requires org.apache.arrow.memory.core;
+  requires org.apache.arrow.vector;
+  requires protobuf.java;
+}
